+++ app/src/main/java/com/yoonbae/planting/planner/data/PlantEvent.java
@@ -5,19 +5,19 @@ import com.prolificinteractive.materialcalendarview.CalendarDay;
 import java.time.LocalTime;
 
 public class PlantEvent {
-    private Long id;
+    private Integer id;
     private String name;
     private String alarmMessage;
     private CalendarDay calendarDay;
 
-    public PlantEvent(Long id, String name, LocalTime alarmTime, CalendarDay calendarDay) {
+    public PlantEvent(Integer id, String name, LocalTime alarmTime, CalendarDay calendarDay) {
         this.id = id;
         this.name = name;
         this.alarmMessage = alarmTime + " 물주기 알람";
         this.calendarDay = calendarDay;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
